The server is finished along side a client
The server has been tested with at least four clients on the same chat thread. It handles forcible disconnection and all the specified types in the message type enum. A scripted version of a client has been written that provides the main use cases: connecting, registering, joining a thread, sending a message, disconnecting. Batch files have been provided for the compilation and execution of both client and server.